test: chatList 조회 Auth.tsx -> ChatList.tsx
- ChatList.tsx로 socket on/off, emit(sub/unsub)  이동

#67
