+++ src/pages/auth/Auth.tsx
@@ -27,30 +27,12 @@ const GameRoom = loadable(() => {
 
 function Auth() {
   const [profileUser, setProfileUser] = useState(getUsername());
-  const [chatList, setChatList] = useState<ChatListType[]>([]);
-  const [myChatList, setMyChatList] = useState<ChatListType[]>([]);
   const socket = getSocket();
 
-  const chatRoomListListener = (res: ChatRoomListType) => {
-    if (res.type === "otherRoom") {
-      const tmp: ChatListType[] = [];
-      res.list.map((elem) => {
-        if (elem.status !== "private") {
-          tmp.push(elem);
-        }
-      });
-      setChatList(tmp);
-    } else if (res.type === "myRoom") {
-      setMyChatList(res.list);
-    }
-  };
-
   useEffect(() => {
-    socket.on("message", chatRoomListListener);
     socket.on("subscribeResult", (data) => console.log(data));
     socket.on("unsubscribeResult", (data) => console.log(data));
     return () => {
-      socket.off("message", chatRoomListListener);
       socket.off("subscribeResult", (data) => console.log(data));
       socket.off("unsubscribeResult", (data) => console.log(data));
     };
@@ -67,7 +49,7 @@ function Auth() {
             <ListTabBar />
             <Routes>
               <Route path="/" element={<Main />} />
-              <Route path="/chat/list" element={<ChatList chat={chatList} myChat={myChatList} />} />
+              <Route path="/chat/list" element={<ChatList />} />
               <Route path="/game/list" element={<GameList />} />
               <Route path="/chat/:roomId" element={<ChatRoom />} />
               <Route path="/game/:gameId" element={<GameRoom />} />
